# api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Jobberwocky",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://localhost:251"
        }
    ],
    "paths": {
...
        "/api/offers": {
            "get": {
                "tags": [
                    "Offer"
                ],
                "summary": "Show all offers of jobs",
                "operationId": "f44a8eb32eda1fa0dc80455c835746b2",
                "responses": {
                    "200": {
                        "description": "ok."
                    },
                    "default": {
                        "description": "An error has occurred."
                    }
                }
            }
        },
        "/api/showOffer/{id}": {
            "get": {
                "tags": [
                    "Offer"
                ],
                "summary": "Displays the details of the offer",
                "operationId": "a77ab4e8cfe1281f5a90da7910529243",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Offer identifier",
                        "required": true,
                        "explode": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok."
                    },
                    "404": {
                        "description": "404 Not Found."
                    }
                }
            }
        },
        "/api/searchOffer/{query}/{country_id}": {
            "get": {
                "tags": [
                    "Offer"
                ],
                "summary": "Displays the details of the offer",
                "operationId": "c06e987b9c59e91f65ab71b2dab42351",
                "parameters": [
                    {
                        "name": "query",
                        "in": "path",
                        "description": "Name of the offer",
                        "required": false,
                        "explode": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "country_id",
                        "in": "path",
                        "description": "Country identifier",
                        "required": true,
                        "explode": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok."
                    },
                    "404": {
                        "description": "404 Not Found."
                    }
                }
            }
        },
        "/api/storeOffer": {
            "post": {
                "tags": [
                    "Offer"
                ],
                "summary": "Add a new offer fito",
                "operationId": "bd2c9e0cefe8e58344eeffeb926ea3e3",
                "requestBody": {
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "name of the new job offer",
                                        "format": "string",
                                        "default": "New job offer"
                                    },
                                    "description": {
                                        "description": "Description of the new job offer",
                                        "format": "string",
                                        "default": "Description of the new job offer"
                                    },
                                    "remote": {
                                        "description": "remote job",
                                        "format": "boolean",
                                        "default": "1"
                                    },
                                    "salary": {
                                        "description": "Salary of the job",
                                        "format": "number",
                                        "default": "42,000 $"
                                    },
                                    "country_id": {
                                        "description": "country id",
                                        "format": "number",
                                        "default": "1"
                                    },
                                    "skills": {
                                        "description": "id skills separated by comma",
                                        "format": "string",
                                        "default": "5, 2"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "ok."
                    },
                    "404": {
                        "description": "404 Not Found."
                    }
                }
            }
        },
        "/api/updateOffer": {
            "put": {
                "tags": [
                    "Offer"
                ],
                "summary": "Upgrade an existing offer.",
                "operationId": "68d6d9b367d3f143f7458a71d9d7d9ce",
                "requestBody": {
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "properties": {
                                    "id": {
                                        "description": "id",
                                        "format": "id",
                                        "default": "1"
                                    },
                                    "name": {
                                        "description": "name of the new job offer",
                                        "format": "string",
                                        "default": "New job offer"
                                    },
                                    "description": {
                                        "description": "Description of the new job offer",
                                        "format": "string",
                                        "default": "Description of the new job offer"
                                    },
                                    "remote": {
                                        "description": "remote job",
                                        "format": "boolean",
                                        "default": "1"
                                    },
                                    "salary": {
                                        "description": "Salary of the job",
                                        "format": "number",
                                        "default": "42,000 $"
                                    },
                                    "country_id": {
                                        "description": "country id",
                                        "format": "number",
                                        "default": "1"
                                    },
                                    "skills": {
                                        "description": "skills",
                                        "format": "string",
                                        "default": "5, 2"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "ok."
                    },
                    "404": {
                        "description": "404 Not Found."
                    }
                }
            }
        },
        "/api/delOffer/{id}": {
            "delete": {
                "tags": [
                    "Offer"
                ],
                "summary": "Delete a selected offer",
                "operationId": "877ccd3ac6d89dc004135f102dd3bb47",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Offer identifier",
                        "required": true,
                        "explode": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok."
                    },
                    "404": {
                        "description": "404 Not Found."
                    }
                }
            }
        },
...
    },
    "security": [
        []
    ]
}